+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Head
 
+* Added: `selector-id-pattern` rule.
 * Added: `function-parentheses-newline-inside` rule.
 * Added: `"always-single-line"` and `"never-single-line"` options to `function-parentheses-space-inside`.